Extract gpu administration task into additional structure (#17)
This PR restructures the lib a bit to separate attestation part and gpu
administration part. Originally all interactions with GPU was connected
with attestation: collect evidence and validate evidence. Now that we
add a few more interactions that are indirectly required for
attestation, we have to extract gpu administration as a standalone unit.

Changes:
* Splits attester into RemoteAttester and GPUAdmin
* Adds a few GPU control methods to GPUAdmin